[SPOT-280][REFACTOR] Organize ToDo Domain
