+++ src/main/java/com/example/spot/refactor/study/application/MemberStudyCommandServiceImpl.java
@@ -13,7 +13,7 @@ import com.example.spot.refactor.schedule.domain.aggregate.StudyQuizSubmission;
 import com.example.spot.refactor.schedule.domain.repository.StudyQuizRepository;
 import com.example.spot.refactor.schedule.domain.repository.StudyQuizSubmissionRepository;
 import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDo;
+import com.example.spot.refactor.todo.domain.StudyToDo;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.legacy.domain.enums.NotifyType;
@@ -30,7 +30,7 @@ import com.example.spot.legacy.repository.NotificationRepository;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostReportRepository;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDoRepository;
+import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.study.presentation.dto.request.ScheduleRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyHostWithdrawRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyMemberReportDTO;
