+++ src/main/java/com/example/spot/refactor/study/application/MemberStudyQueryServiceImpl.java
@@ -16,13 +16,13 @@ import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.schedule.domain.enums.StudySchedulePeriod;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDo;
+import com.example.spot.refactor.todo.domain.StudyToDo;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVote;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDoRepository;
+import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.study.presentation.dto.response.ScheduleResponseDTO;
 import com.example.spot.refactor.study.presentation.dto.response.StudyImageResponseDTO;
 import com.example.spot.refactor.study.presentation.dto.response.StudyMemberResDTO;
