+++ src/main/java/com/example/spot/refactor/todo/domain/StudyToDo.java
@@ -1,4 +1,4 @@
-package com.example.spot.refactor.study.domain.aggregate.studytodo;
+package com.example.spot.refactor.todo.domain;
 
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.common.entity.BaseEntity;
