+++ src/test/java/com/example/spot/service/study/studymember/StudyMemberQueryServiceTest.java
@@ -4,7 +4,7 @@ import com.example.spot.refactor.common.api.exception.GeneralException;
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.schedule.domain.StudySchedule;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDo;
+import com.example.spot.refactor.todo.domain.StudyToDo;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
 import com.example.spot.refactor.study.domain.aggregate.Study;
@@ -13,7 +13,7 @@ import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDoRepository;
+import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.common.security.utils.SecurityUtils;
 import com.example.spot.refactor.study.application.MemberStudyQueryServiceImpl;
 import com.example.spot.refactor.study.presentation.dto.response.ToDoListResponseDTO.ToDoListSearchResponseDTO;
